Release v2.0.0 (#122)
* [CSI-1150] - SDK - Conversation.getEvents events has nil member

* Fixed a bug which prevented getting onMemberEvent from push notfications. Description: tryUpdateConversationSequenceId was called twice in the same flow, hence the second call in OnMemberEvent failed. Fix: remove first call of tryUpdateConversatioSequenceId from proccesNexmoPushWithUserInfo (#115)

tryUpdateConversationSequenceId was called twice in the same flow, hence the second call in OnMemberEvent failed. Fix: remove first call of tryUpdateConversatioSequenceId from proccesNexmoPushWithUserInfo

* [CSI-1149]  - SDK - not notify error on bad ACL (system:error:permission)

* [NO-VAR] update changelog and version to 1.2.3


* [CSI-1151] - SDK - Update user agent format: `NexmoClientSDK/[SDK_VER] [PLATFORM] ([ORIGINAL_SDK])`

* [NO-VAR] update changelog and version to 1.2.4

* [CSI-1150] - SDK - Conversation.getEvents for memberEvent use memberId field instead of fromMemberId

* [CSI-1150] - SDK - Conversation.getEvents for memberEvent use memberId to update member object

* [CSI-1150] - SDK - fix for Conversation.getEvents for memberEvent use memberId to update member object

* [NO-VAR] update changelog and version to 2.0.0

* [NO-VAR] update readme and contribute guidelines

* [NO-VAR] - readme fix

* [NO-VAR] - update contribute.md

* [NO-VAR] - update readme

* update diagram

* [CSI-1155] - TestApp - Fix login after logout (#118)

* added socket connection method when login with button
* all login methods now connect to socket immediately
* extracted logic related to login from init method and created loginWithUserToken method

* [CSI-1156] SDK - clear warnings, update mini and infra frameworks versions (#119)

* [CSI-1156] - TestApp - clear logs

* [NO-VAR] update CONTRIBUTE.md

* [NO-VAR] update README.md

* [CSI-1138] - SDK- Get events page module and events filtering (#117)

* [NO-VAR] - TestApp - remove certificate from the repo, moved to nexmoclient-secrets

* [NO-VAR] - TestApp - EventsTableVC update default page size and add event.description to event cell

* [CSI-1159] - SDK - get conversations in parallel (#120)

* getConversationsFromIds now works in parallel
* method signature was changed (now accepts only completion handler block)

* CSI-1168 - Get events page: removing implicit unknown events filtering (#121)

* [NO VAR] -  SDK - Version bumped to 2.1.0 (#123)

+++ NexmoClient/Internal/NexmoCore/Utility/NXMErrorParser.m
@@ -41,7 +41,7 @@ static NSDictionary<NSString *,NSNumber *> *csErrorToNXMCode;
 + (int) parseError:(nonnull NSDictionary*) data{
     
     NSString* errorCodeMsg = data[@"code"];
-    return csErrorToNXMCode[errorCodeMsg] ? csErrorToNXMCode[errorCodeMsg].integerValue : NXMErrorCodeUnknown;
+    return csErrorToNXMCode[errorCodeMsg] ? (int)csErrorToNXMCode[errorCodeMsg].integerValue : (int)NXMErrorCodeUnknown;
 }
 
 @end
